requirements.txt
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.3
google-auth==1.6.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kubernetes==9.0.0
MarkupSafe==1.1.1
oauthlib==3.0.1
prometheus-client==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-dateutil==2.8.0
PyYAML==5.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
urllib3==1.25.3
websocket-client==0.56.0
Werkzeug==0.15.4
